# backend/docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted IoT platform backend. Two binaries: api (HTTP/WS) and ingest (MQTT).",
        "title": "IoT Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "IoT Platform Team",
            "email": "dev@iotplatform.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.AuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal.APIError"
                        }
                    }
                }
            }
        },
...
        "/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh JWT tokens",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal.APIError"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal.APIError"
                        }
                    }
                }
            }
        },
...
        "/devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "List user's devices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal.Device"
                            }
                        }
                    }
                }
            }
        },
        "/devices/{key}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get device details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.Device"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal.APIError"
                        }
                    }
                }
            }
        },
        "/devices/{key}/claim": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Claim an unclaimed device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Device API key",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.ClaimDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal.APIError"
                        }
                    }
                }
            }
        },
...
        "/telemetry/{key}/latest": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Telemetry"
                ],
                "summary": "Get latest telemetry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device key",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.LatestTelemetry"
                        }
                    }
                }
            }
        },
        "/users/me/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Get notification preferences",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal.NotificationPrefs"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Update notification preferences",
                "parameters": [
                    {
                        "description": "Preferences",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.UpdateNotificationPrefsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/internal.APIErrorDetail"
                }
            }
        },
        "internal.APIErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "internal.AuthResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/internal.User"
                }
            }
        },
        "internal.ClaimDeviceRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                }
            }
        },
...
        "internal.Device": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "device_key": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "device_type": {
                    "type": "string"
                },
                "firmware_ver": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "owner_id": {
                    "description": "nil = unclaimed",
                    "type": "string"
                }
            }
        },
...
        "internal.LatestTelemetry": {
            "type": "object",
            "properties": {
                "battery_power": {
                    "type": "number"
                },
                "dc_load_power": {
                    "type": "number"
                },
                "device_key": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "inverter_power": {
                    "type": "number"
                },
                "max_soc_pct": {
                    "type": "number"
                },
                "min_soc_pct": {
                    "type": "number"
                },
                "pv_power": {
                    "type": "number"
                },
                "recorded_at": {
                    "type": "string"
                },
                "system_status": {
                    "type": "integer"
                },
                "total_energy_wh": {
                    "type": "number"
                }
            }
        },
        "internal.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "internal.NotificationPrefs": {
            "type": "object",
            "properties": {
                "alert_fired_email": {
                    "type": "boolean"
                },
                "alert_resolved_email": {
                    "type": "boolean"
                },
                "quiet_hours_end": {
                    "type": "integer"
                },
                "quiet_hours_start": {
                    "type": "integer"
                }
            }
        },
...
        "internal.RefreshRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "internal.RegisterRequest": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "internal.UpdateNotificationPrefsRequest": {
            "type": "object",
            "properties": {
                "alert_fired_email": {
                    "type": "boolean"
                },
                "alert_resolved_email": {
                    "type": "boolean"
                },
                "quiet_hours_end": {
                    "type": "integer"
                },
                "quiet_hours_start": {
                    "type": "integer"
                }
            }
        },
        "internal.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token from /auth/login or /auth/register",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}